% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genr.R
\name{Gen.r}
\alias{Gen.r}
\title{Generate data with a specific sample correlation.}
\usage{
Gen.r(x, r)
}
\arguments{
\item{x}{a vector of values}

\item{r}{the sample correlation}
}
\description{
This function allows you to generate data with a specific sample correlation to another vector.
}
\keyword{Gen.r()}
\keyword{correlation}
\keyword{data}
\keyword{generation}
