% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncrementFunction.R
\name{i}
\alias{i}
\title{Increment Function}
\usage{
i(x, v = 1, envir = parent.frame())
}
\arguments{
\item{x}{The variable name you want to increment.}

\item{v}{Value to increment x by. Does not necesssarily need to be an integer. Default = 1.}
}
\description{
This function allows you to increment a variable by 1, or a specified value. If the variable name does not exist, the function creates it and sets the value to one.
}
\keyword{Increment}
\keyword{i()}
