% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function to check whether package is installed
%   and install if not.R
\name{InstLoad}
\alias{InstLoad}
\title{Load or Install then load a package}
\usage{
InstLoad(pkg)
}
\arguments{
\item{pgk}{The}
}
\description{
Loads a library. In the library is not installed, the function will install then load the package.
}
\keyword{InstLoad()}
\keyword{R}
\keyword{libraries}
\keyword{packages,}
