% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-fold cross validation.R
\name{Kfold}
\alias{Kfold}
\title{K-fold Cross Validation}
\usage{
Kfold(dat, k)
}
\arguments{
\item{dat}{The dataset to split}

\item{k}{The number of folds
Kfold()}
}
\description{
This function splits the dataset into K-folds with a test set and a train set, returning the split datasets as a list.
}
