% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolychoricMatrix.R
\name{Polycor.matrix}
\alias{Polycor.matrix}
\title{Calculate Polycoric Correlation Matrices using multiple cores}
\usage{
Polycor.matrix(dat, n.cores = 1)
}
\arguments{
\item{dat}{A dataset of ordered categorical data}

\item{n.cores}{The number of cores to use. See function detectCores(). default = 1.}
}
\description{
This function estimates a pairwise polychoric correlation matrix using the mclapply function in the paralell packages. Note windows machines are unable to do paralell processing using forking. See package snow for more options.
}
\keyword{Correlation}
\keyword{Matrix}
\keyword{Polychoric}
\keyword{Polycor.matrix()}
