% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RpbtoRbis.R
\name{RpbtoRbis}
\alias{RpbtoRbis}
\title{Convert a point-biserial correlation to a biserial correlation}
\usage{
RpbtoRbis(Rpb, p)
}
\arguments{
\item{Rpb}{A point-biserial correlation}

\item{p}{The proportion of "1's" in the dichotomous variable}
}
\description{
This function allows you to to convert a point-biserial correlation to a biserial correlation
}
\keyword{Biserial}
\keyword{RpbtoRbis()}
\keyword{correlation,}
\keyword{point-biserial}
