% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taylor-Russell.R
\name{TaylorRussell}
\alias{TaylorRussell}
\title{Taylor Russell Tables}
\usage{
TaylorRussell(rho, base, select)
}
\arguments{
\item{rho}{The validity of the assessment}

\item{base}{The base rate of potential hires in the population that would be successful}

\item{select}{The selection ratio: i.e. the top x% of employees to be selected.}
}
\description{
Taylor Russell tables estimate the percentage of new hires that will be successful if a particular selection method is used.
}
\keyword{Employee}
\keyword{Selection}
\keyword{TaylorRussell()}
