% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt.R
\name{TestRetest}
\alias{TestRetest}
\title{Test Retest Matrix}
\usage{
TestRetest(dat1, dat2, use = c("everything", "all.obs", "complete.obs",
  "na.or.complete", "pairwise.complete.obs"))
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Creates a matrix of correlations with inter-item correlations from time 1 and time 2 on the respective off diagonals with test retest reliability on the main diagonal. Useful in conjunction with corrplot().
}
\keyword{TestRetest()}
\keyword{cats}
