% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taylor-Russell.R
\name{TaylorRussell}
\alias{TaylorRussell}
\title{Taylor Russell Tables}
\usage{
TaylorRussell(rho, base, select)
}
\arguments{
\item{rho}{A vector of validity values for an assessment}

\item{base}{A vector of base rates of potential hires in the population that would be successful}

\item{select}{A vector of selction ratios: i.e. the top x% of employees to be selected.}
}
\description{
Taylor Russell tables estimate the percentage of new hires that will be successful if a particular selection method is used.
}
\keyword{Russell}
\keyword{Taylor}
\keyword{TaylorRussell()}
